step3-review: CarRacingTest 와 CarTest 분리
